+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/AcademiesDbServices/GetEstablishments.cs
@@ -1,8 +1,9 @@
 using System.Net.Http.Json;
 using System.Web;
+using Dfe.AcademiesApi.Client.Contracts;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Http;
-using GovUK.Dfe.CoreLibs.Contracts.Academies.V4.Establishments;
 using Microsoft.Extensions.Logging;
+using EstablishmentDto = GovUK.Dfe.CoreLibs.Contracts.Academies.V4.Establishments.EstablishmentDto;
 
 
 namespace DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.AcademiesDbServices;
@@ -28,6 +29,17 @@ public class GetEstablishments(IDfeHttpClientFactory httpClientFactory,
         string path = $"v4/establishment/urn/{urn}";
 
         ApiResponse<EstablishmentDto> result = await httpClientService.Get<EstablishmentDto>(_httpClient, path);
+
+        if (!result.Success) throw new ApiResponseException($"Request to Api failed | StatusCode - {result.StatusCode}");
+
+        return result.Body;
+    }
+
+    public async Task<EstablishmentResponse> GetEstablishmentWithSenData(int urn)
+    {
+        string path = $"establishment/urn/{urn}";
+
+        ApiResponse<EstablishmentResponse> result = await httpClientService.Get<EstablishmentResponse>(_httpClient, path);
 
         if (!result.Success) throw new ApiResponseException($"Request to Api failed | StatusCode - {result.StatusCode}");