+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/Repositories/SchoolRepository.cs
@@ -31,30 +31,26 @@ public class SchoolRepository(IAcademiesDbContext academiesDbContext,
 
     public async Task<SchoolDetails> GetSchoolDetailsAsync(int urn)
     {
-        return await academiesDbContext.GiasEstablishments
-            .Where(e => e.Urn == urn)
-            .Select(establishment => new SchoolDetails(establishment.EstablishmentName!,
-                stringFormattingUtilities.BuildAddressString(
-                    establishment.Street,
-                    null,
-                    establishment.Town,
-                    establishment.Postcode
-                ),
-                establishment.GorName!,
-                establishment.LaName!,
-                establishment.PhaseOfEducationName!,
-                new AgeRange(establishment.StatutoryLowAge, establishment.StatutoryHighAge),
-                establishment.NurseryProvisionName))
-            .SingleAsync();
-    }
+        var result = await getEstablishments.GetEstablishment(urn);
+        
+        var dateJoinedTrust = !string.IsNullOrEmpty(result.DateJoinedTrust) ? DateTime.Parse(result.DateJoinedTrust, CultureInfo.InvariantCulture) : (DateTime?)null;
 
-    public async Task<DateOnly?> GetDateJoinedTrustAsync(int urn)
-    {
-        return await academiesDbContext.GiasGroupLinks.Where(gl => gl.Urn == urn.ToString())
-            .Select(gl =>
-                DateOnly.ParseExact(gl.JoinedDate!, "dd/MM/yyyy", CultureInfo.InvariantCulture, DateTimeStyles.None))
-            .Cast<DateOnly?>()
-            .FirstOrDefaultAsync();
+
+        return new SchoolDetails(
+            Name:  result.Name,
+                Address: stringFormattingUtilities.BuildAddressString(
+                    result.Address.Street,
+                    result.Address.Locality,
+                    result.Address.Town,
+                    result.Address.Postcode
+                ),
+                Region: result.Gor.Name,
+                LocalAuthority: result.LocalAuthorityName,
+                PhaseOfEducationName: result.PhaseOfEducation.Name,
+                AgeRange: new AgeRange(result.StatutoryLowAge, result.StatutoryHighAge),
+                NurseryProvision: result.NurseryProvision,
+                TrustName: result.TrustName,
+                DateJoinedTrust: dateJoinedTrust);
     }
 
     public async Task<SchoolContact?> GetSchoolContactsAsync(int urn)
@@ -79,32 +75,33 @@ public class SchoolRepository(IAcademiesDbContext academiesDbContext,
 
     public async Task<SenProvision> GetSchoolSenProvisionAsync(int urn)
     {
-        return await academiesDbContext.GiasEstablishments
-            .Where(e => e.Urn == urn)
-            .Select(establishment => new SenProvision(
-                establishment.ResourcedProvisionOnRoll!,
-                establishment.ResourcedProvisionCapacity!,
-                establishment.SenUnitOnRoll!,
-                establishment.SenUnitCapacity!,
-                establishment.TypeOfResourcedProvisionName!,
-                new List<string>
-                {
-                    establishment.Sen1Name!,
-                    establishment.Sen2Name!,
-                    establishment.Sen3Name!,
-                    establishment.Sen4Name!,
-                    establishment.Sen5Name!,
-                    establishment.Sen6Name!,
-                    establishment.Sen7Name!,
-                    establishment.Sen8Name!,
-                    establishment.Sen9Name!,
-                    establishment.Sen10Name!,
-                    establishment.Sen11Name!,
-                    establishment.Sen12Name!,
-                    establishment.Sen13Name!
-                }
-            ))
-            .SingleAsync();
+        var result = await getEstablishments.GetEstablishmentWithSenData(urn);
+        
+        var senProvision = new SenProvision(
+            result.ResourcedProvisionOnRoll,
+            result.ResourcedProvisionOnCapacity,
+            result.SenUnitOnRoll,
+            result.SenUnitCapacity,
+            result.TypeOfResourcedProvision,
+            new List<string>
+            {
+                result.SeN1!,
+                result.SeN2!,
+                result.SeN3!,
+                result.SeN4!,
+                result.SeN5!,
+                result.SeN6!,
+                result.SeN7!,
+                result.SeN8!,
+                result.SeN9!,
+                result.SeN10!,
+                result.SeN11!,
+                result.SeN12!,
+                result.SeN13!
+            }
+        );
+        
+        return senProvision;
     }
 
     public async Task<bool> IsPartOfFederationAsync(int urn)